fet: finish build structure
