+++ src/config.ts
@@ -1,0 +1,24 @@
+import * as path from 'path';
+import * as vscode from 'vscode';
+import * as fs from 'fs';
+
+async function checkAndCreateConfigs() {
+    const workspaceFolder = vscode.workspace.workspaceFolders?.[0]?.uri.fsPath;
+    if (!workspaceFolder) return;
+
+    const tasksPath = path.join(workspaceFolder, '.vscode', 'tasks.json');
+    if (!fs.existsSync(tasksPath)) {
+        const choice = await vscode.window.showInformationMessage(
+            'Create FASM build configuration?',
+            'Yes', 'No'
+        );
+        
+        if (choice === 'Yes') {
+            await vscode.commands.executeCommand('fasm.createConfigs');
+        }
+    }
+}
+
+export default {
+    checkAndCreateConfigs
+}
